pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uavtools"
version = "1.3"
description = "FastMCP UAV Tools - Reconnaissance and Detection Service"
authors = [{ name = "Cloud" }]
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "fastapi-mcp>=0.4.0",
    "fastapi[standard]>=0.117.1",
    "fastmcp==2.10.0",
    "flask>=3.1.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-deepseek>=0.1.4",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "mysql-connector-python>=9.4.0",
    "pymilvus[model]>=2.6.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "python-socketio[asyncio-client]>=5.13.0",
    "pyyaml>=6.0.3",
    "shapely>=2.1.2",
    "timm>=1.0.20",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.56.2",
    "ultralytics>=8.3.203",
    "websockets>=15.0.1",
    "watchdog>=4.0.0",
]

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.setuptools.packages.find]
where = ["."]

# CPU version PyTorch installation:
# After installation, run the following command to ensure CPU version:
# uv pip install torch torchvision --index-url https://mirrors.tuna.tsinghua.edu.cn/pytorch/whl/cpu --upgrade

